backend/pyproject.toml
[project]
name = "vortixpr-backend"
version = "1.0.0"
description = "VortixPR Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.1",
    "asyncpg==0.30.0",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.0",
    "email-validator==2.2.0",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",
    "python-slugify==8.0.4",
    "python-multipart==0.0.20",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []

